# [android] improve performance of `StrokeExtensions`