Context: https://github.com/dotnet/maui/issues/23991
Context: https://github.com/chabiss/periodictable

In the above sample in `Debug`-mode, a lot of time is spent in:

    1.20s (3.3%) microsoft.maui!Microsoft.Maui.Platform.StrokeExtensions.UpdateBorderStroke(Android.Views.View,Microsoft.Maui.IBorderStroke)
    0.42s (1.2%) microsoft.maui!Microsoft.Maui.Platform.StrokeExtensions.UpdateMauiDrawable(Android.Views.View,Microsoft.Maui.IBorderStroke)

If you drill-in, a lot of the time is spent in, `View.Background`, which is called a lot overall as well:

    627.20ms (1.7%) mono.android!Android.Views.View.get_Background()

Reviewing `StrokeExtensions`, it does call `View.Background` over and over, many times.

To fix this:

* Call `View.Background` once and store it in a `MauiDrawable?` local variable.

* Subsequent calls to methods in `StrokeExtensions` pass the local variable along.

`View.Background` is now called many fewer times and the result is:

    1.07s (2.90%) microsoft.maui!Microsoft.Maui.Platform.StrokeExtensions.UpdateBorderStroke(Android.Views.View,Microsoft.Maui.IBorderStroke)
    0.34s (0.94%) microsoft.maui!Microsoft.Maui.Platform.StrokeExtensions.UpdateMauiDrawable(Android.Views.View,Microsoft.Maui.IBorderStroke,Microsoft.Maui.Graphics.MauiDrawable&)
    0.33s (0.91%) mono.android!Android.Views.View.get_Background()

This should improve the performance of controls that use `MauiDrawable` like `<Border/>` on Android.

For the future, we're looking into other ways to improve Debug-mode performance at:

* https://github.com/dotnet/runtime/issues/106491